COL-1967, whiteboard: /upserts via socktes, not http API
